+++ testutils/src/main/java/com/google/android/exoplayer2/testutil/StubExoPlayer.java
@@ -131,6 +131,11 @@ public abstract class StubExoPlayer implements ExoPlayer {
   }
 
   @Override
+  public void stop(boolean resetStateAndPosition) {
+    throw new UnsupportedOperationException();
+  }
+
+  @Override
   public void release() {
     throw new UnsupportedOperationException();
   }